Test That a Player Wearing the Black Helmet is not attacked by I.D.
This test is currently failing as the feature has not yet been
implemented. This test places a player who is wearing the black helmet
in a zone with an imperial defender, and sets the player as the target
of the imperial defender. It asserts that the HP of the player is not
changed by this.