Fully Implement the Enchanted Black Helmet as a Disguise
Tests have been edited to check for invisibility being disabled when the
item is taken off because the previous testing of for a change in HP was
failing despite the player becoming invisible during game play. I have
also added the onUnequipped method to he EnchantedBlackHelmet class so
that the player becomes 'visible' again when the helmet is taken off.
The item has been added to helmets.xml so that it is available by admin
summon. The SummonActionTest has also been updated to check that the
enchanted black helmet is summoned, rather than the black helmet.